rewrite the doctest reformat tests (#138)
The tests were not particularly easy to write, nor to read and understand. This means that the test coverage is less than it could be, which in turn allowed several bugs to get introduced in the last release (see #137).

<!-- Feel free to remove check-list items aren't relevant to your change -->

 - [x] Tests added
 - [x] Passes `pre-commit run --all-files`

<!--
By default, the upstream-dev is only run when triggered by the github website (`workflow_dispatch`)
or if it was run on a schedule. To run it on a commit of a pull request (`pull_request`), include
the `[test-upstream]` tag in the summary line of the commit message.

For changes that are not covered by the CI please use the `[skip-ci]` tag to avoid running the
normal CI.
-->